quiz_flask/requirements.txt
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
diskcache==5.6.3
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
json_repair==0.54.3
llama_cpp_python==0.3.16
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.0
packaging==25.0
py-cpuinfo==9.0.0
python-docx==1.2.0
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
tqdm==4.67.1
typer-slim==0.21.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
